--- Muse/Assets/scripts/DeWKnop.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class DeWKnop : MonoBehaviour
-{
-    public bool grav = false;
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
